Work on enter/leave and focus handling
