+++ mfp/gui/clutter/app_window.py
@@ -557,14 +557,13 @@ class ClutterAppWindowBackend (AppWindowBackend):
             else:
                 element.container = element.layer.backend.group
 
-        self.wrapper.input_mgr.backend.event_source_reverse[element.group] = element
+        self.wrapper.event_sources[element.group] = element
 
         if not isinstance(element, ConnectionElement):
             if self.wrapper.load_in_progress:
                 update = False
             else:
                 update = True
-
 
     def unregister(self, element):
         if element.container:
@@ -573,8 +572,8 @@ class ClutterAppWindowBackend (AppWindowBackend):
             element.container = None
 
         self.object_view.remove(element)
-        if element.group in self.wrapper.input_mgr.backend.event_source_reverse:
-           del self.wrapper.input_mgr.backend.event_source_reverse[element.group]
+        if element.group in self.wrapper.event_sources:
+           del self.wrapper.event_sources[element.group]
 
     def refresh(self, element):
         if isinstance(element, PatchDisplay):
